+++ style/markdown.go
@@ -1,5 +1,10 @@
 package style
 
+import (
+	"fmt"
+	"github.com/pirmd/cli/style/text"
+)
+
 //Markdown is a sub-set of markdown markup
 var Markdown = PlainText.Extend(New(
 	FormatMap{
@@ -14,6 +19,15 @@ var Markdown = PlainText.Extend(New(
 		FmtCode: Sprintf("`%s`"),
 	},
 	nil,
+	nil,
+	nil,
+	nil,
+
+	//defineFn
+	func(term, desc string) string {
+		s := fmt.Sprintf("\n%s\n:%s\n", term, desc)
+		return text.Wrap(s, DefaultTxtWidth)
+	},
 ))
 
 //XXX: Introduce Code and Bloc, transfer them to plaintext?